refactor: implement events list
